feat(ui): extract EntityPureUtils, EntityBreadcrumbUtils, and EntityLabel from EntityUtils
Split EntityUtils.tsx (1574 lines) into focused modules:
- EntityPureUtils.ts: 24 pure functions (no JSX, no side effects)
- EntityBreadcrumbUtils.tsx: all breadcrumb and entity link functions
- EntityLabel.component.tsx: proper React component extracted from getEntityLabel
- EntityUtils.tsx: 183 lines — re-exports for backward compat + 2 JSX renderers

All original exports remain available from EntityUtils via re-exports.
