+++ openmetadata-ui/src/main/resources/ui/src/utils/EntityPureUtils.ts
@@ -1,0 +1,397 @@
+/*
+ *  Copyright 2025 Collate.
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+import { isEmpty, isUndefined, lowerCase } from 'lodash';
+import { EntityDetailUnion } from 'Models';
+import { Node } from 'reactflow';
+import { QueryVoteType } from '../components/Database/TableQueries/TableQueries.interface';
+import {
+  CUSTOM_PROPERTIES_TABS_SET,
+  LINEAGE_TABS_SET,
+  SCHEMA_TABS_SET,
+} from '../components/Entity/EntityRightPanel/EntityRightPanelVerticalNav.constants';
+import { SearchedDataProps } from '../components/SearchedData/SearchedData.interface';
+import {
+  PLACEHOLDER_ROUTE_ENTITY_TYPE,
+  PLACEHOLDER_ROUTE_FQN,
+  ROUTES,
+} from '../constants/constants';
+import { EntityLineageNodeType, EntityType } from '../enums/entity.enum';
+import {
+  SearchIndexField,
+} from '../generated/entity/data/searchIndex';
+import {
+  Column,
+  ColumnJoins,
+  JoinedWith,
+  Table,
+} from '../generated/entity/data/table';
+import { User } from '../generated/entity/teams/user';
+import { EntityReference } from '../generated/type/entityUsage';
+import { TagLabel } from '../generated/type/tagLabel';
+import { Votes } from '../generated/type/votes';
+import EntityLink from './EntityLink';
+import { getEntityName } from './EntityNameUtils';
+import Fqn from './Fqn';
+import { getEncodedFqn } from './StringUtils';
+import { getDataTypeString, getTagsWithoutTier } from './TableUtils';
+import { getTableTags } from './TagsUtils';
+
+export enum DRAWER_NAVIGATION_OPTIONS {
+  explore = 'Explore',
+  lineage = 'Lineage',
+}
+
+export const ENTITY_LINK_SEPARATOR = '::';
+
+export const getEntityFeedLink = (
+  type?: string,
+  fqn?: string,
+  field?: string
+): string => {
+  if (isUndefined(type) || isUndefined(fqn)) {
+    return '';
+  }
+
+  return `<#E${ENTITY_LINK_SEPARATOR}${type}${ENTITY_LINK_SEPARATOR}${fqn}${
+    field ? `${ENTITY_LINK_SEPARATOR}${field}` : ''
+  }>`;
+};
+
+/*
+  params: userName - fullyQualifiedName
+  return : <#E::user::userName>
+*/
+export const getEntityUserLink = (userName: string): string => {
+  return `<#E${ENTITY_LINK_SEPARATOR}user${ENTITY_LINK_SEPARATOR}${userName}>`;
+};
+
+export const checkIfJoinsAvailable = (
+  columnName: string,
+  joins: Array<ColumnJoins>
+): boolean => {
+  return (
+    joins &&
+    Boolean(joins.length) &&
+    Boolean(joins.some((join) => join.columnName === columnName))
+  );
+};
+
+/**
+ * It takes a column name and a list of joins and returns the list of joinedWith for the column name
+ * @param {string} columnName - The name of the column you want to get the frequently joined with
+ * columns for.
+ * @param joins - Array<ColumnJoins>
+ * @returns An array of joinedWith objects
+ */
+export const getFrequentlyJoinedWithColumns = (
+  columnName: string,
+  joins: Array<ColumnJoins>
+): Array<JoinedWith> => {
+  return (
+    (joins &&
+      Boolean(joins.length) &&
+      joins?.find((join) => join.columnName === columnName)?.joinedWith) ||
+    []
+  );
+};
+
+export const getEntityReferenceFromEntity = <
+  T extends Omit<EntityReference, 'type'>
+>(
+  entity: T,
+  type: EntityType
+): EntityReference => {
+  return {
+    id: entity.id,
+    type,
+    deleted: entity.deleted,
+    description: entity.description,
+    displayName: entity.displayName,
+    fullyQualifiedName: entity.fullyQualifiedName,
+    href: entity.href,
+    name: entity.name,
+  };
+};
+
+/**
+ * Convert all the entity list to EntityReferenceList
+ * @param entities -- T extends EntityReference
+ * @param type -- EntityType
+ * @returns EntityReference[]
+ */
+export const getEntityReferenceListFromEntities = <
+  T extends Omit<EntityReference, 'type'>
+>(
+  entities: T[],
+  type: EntityType
+) => {
+  if (isEmpty(entities)) {
+    return [] as EntityReference[];
+  }
+
+  return entities.map((entity) => getEntityReferenceFromEntity(entity, type));
+};
+
+/**
+ * Take entity vote and userId as input and return name for vote status type
+ * @param votes - entity votes
+ * @param userId - current user id
+ * @returns - vote status type
+ */
+export const getEntityVoteStatus = (userId: string, votes?: Votes) => {
+  if (isUndefined(votes)) {
+    return QueryVoteType.unVoted;
+  }
+
+  const upVoters = votes.upVoters ?? [];
+  const downVoters = votes.downVoters ?? [];
+
+  if (upVoters.some((user) => user.id === userId)) {
+    return QueryVoteType.votedUp;
+  } else if (downVoters.some((user) => user.id === userId)) {
+    return QueryVoteType.votedDown;
+  } else {
+    return QueryVoteType.unVoted;
+  }
+};
+
+export const highlightEntityNameAndDescription = (
+  entity: SearchedDataProps['data'][number]['_source'],
+  highlight: SearchedDataProps['data'][number]['highlight']
+): SearchedDataProps['data'][number]['_source'] => {
+  let entityDescription = entity.description ?? '';
+  const descHighlights = highlight?.description ?? [];
+
+  if (descHighlights.length > 0) {
+    const matchTextArr = descHighlights.map((val: string) =>
+      val.replace(/<\/?span(.*?)>/g, '')
+    );
+
+    matchTextArr.forEach((text: string, i: number) => {
+      entityDescription = entityDescription.replace(text, descHighlights[i]);
+    });
+  }
+
+  let entityDisplayName = getEntityName(entity);
+  if (!isUndefined(highlight)) {
+    entityDisplayName =
+      highlight?.displayName?.join(' ') ||
+      highlight?.name?.join(' ') ||
+      entityDisplayName;
+  }
+
+  return {
+    ...entity,
+    displayName: entityDisplayName,
+    description: entityDescription,
+  };
+};
+
+export const columnSorter = (
+  col1: { name: string; displayName?: string },
+  col2: { name: string; displayName?: string }
+) => {
+  const name1 = getEntityName(col1);
+  const name2 = getEntityName(col2);
+
+  return name1.localeCompare(name2);
+};
+
+/**
+ * Retrieves the column name from an entity link.
+ * @param entityLink The entity link string.
+ * @returns The column name extracted from the entity link.
+ */
+export const getColumnNameFromEntityLink = (entityLink: string) => {
+  return EntityLink.getTableColumnName(entityLink);
+};
+
+export const getColumnSorter = <T, K extends keyof T>(field: K) => {
+  return (a: T, b: T) => {
+    const aValue = a[field];
+    const bValue = b[field];
+    if (typeof aValue === 'string' && typeof bValue === 'string') {
+      return aValue.localeCompare(bValue);
+    }
+
+    return 0;
+  };
+};
+
+export const highlightSearchText = (
+  text?: string,
+  searchText?: string
+): string => {
+  if (!searchText || !text) {
+    return text ?? '';
+  }
+
+  const regex = new RegExp(`(${searchText})`, 'gi');
+
+  return text.replace(
+    regex,
+    `<span data-highlight="true" class="text-highlighter">$1</span>`
+  );
+};
+
+export const getEntityImportPath = (entityType: EntityType, fqn: string) => {
+  return ROUTES.ENTITY_IMPORT.replace(
+    PLACEHOLDER_ROUTE_ENTITY_TYPE,
+    entityType
+  ).replace(PLACEHOLDER_ROUTE_FQN, getEncodedFqn(fqn));
+};
+
+export const getEntityBulkEditPath = (entityType: EntityType, fqn: string) => {
+  return ROUTES.BULK_EDIT_ENTITY_WITH_FQN.replace(
+    PLACEHOLDER_ROUTE_ENTITY_TYPE,
+    entityType
+  ).replace(PLACEHOLDER_ROUTE_FQN, getEncodedFqn(fqn));
+};
+
+export const getBreadcrumbsFromFqn = (fqn: string, includeCurrent = false) => {
+  const fqnList = Fqn.split(fqn);
+  if (!includeCurrent) {
+    fqnList.pop();
+  }
+
+  return [
+    ...fqnList.map((fqn) => ({
+      name: fqn,
+      url: '',
+    })),
+  ];
+};
+
+/**
+ * Updates the node type based on whether it's a source or target node
+ * @param node - The node to update
+ * @param sourceNodeId - ID of the source node
+ * @param targetNodeId - ID of the target node
+ * @returns The updated node with the correct type
+ */
+export const updateNodeType = (
+  node: Node,
+  sourceNodeId?: string,
+  targetNodeId?: string
+): Node => {
+  if (node.id === sourceNodeId) {
+    return {
+      ...node,
+      type: EntityLineageNodeType.INPUT,
+    };
+  }
+  if (node.id === targetNodeId) {
+    return {
+      ...node,
+      type: EntityLineageNodeType.OUTPUT,
+    };
+  }
+
+  return node;
+};
+
+export const hasSchemaTab = (entityType: EntityType): boolean =>
+  SCHEMA_TABS_SET.has(entityType);
+
+export const hasLineageTab = (entityType: EntityType): boolean =>
+  LINEAGE_TABS_SET.has(entityType);
+
+export const hasCustomPropertiesTab = (entityType: EntityType): boolean =>
+  CUSTOM_PROPERTIES_TABS_SET.has(entityType);
+
+export const hasEditAccess = (owners: EntityReference[], currentUser: User) => {
+  return owners.some((owner) => {
+    if (owner.type === 'user') {
+      return owner.id === currentUser.id;
+    } else {
+      return Boolean(
+        currentUser.teams?.length &&
+          currentUser.teams.some((team) => team.id === owner.id)
+      );
+    }
+  });
+};
+
+export const getEntityTags = (
+  type: string,
+  entityDetail: EntityDetailUnion
+): Array<TagLabel> => {
+  switch (type) {
+    case EntityType.TABLE: {
+      const tableTags: Array<TagLabel> = [
+        ...getTableTags((entityDetail as Table).columns ?? []),
+        ...(entityDetail.tags ?? []),
+      ];
+
+      return tableTags;
+    }
+    case EntityType.DASHBOARD:
+    case EntityType.SEARCH_INDEX:
+    case EntityType.PIPELINE:
+      return getTagsWithoutTier(entityDetail.tags ?? []);
+
+    case EntityType.TOPIC:
+    case EntityType.MLMODEL:
+    case EntityType.STORED_PROCEDURE:
+    case EntityType.DASHBOARD_DATA_MODEL: {
+      return entityDetail.tags ?? [];
+    }
+
+    default:
+      return [];
+  }
+};
+
+/**
+ * It searches for a given text in a given table and returns a new table with only the columns that
+ * contain the given text
+ * @param {Column[]} table - Column[] - the table to search in
+ * @param {string} searchText - The text to search for.
+ * @returns An array of columns that have been searched for a specific string.
+ */
+export const searchInColumns = <T extends Column | SearchIndexField>(
+  table: T[],
+  searchText: string
+): T[] => {
+  const searchedValue: T[] = table.reduce((searchedCols, column) => {
+    const searchLowerCase = lowerCase(searchText);
+    const isContainData =
+      lowerCase(column.name).includes(searchLowerCase) ||
+      lowerCase(column.displayName).includes(searchLowerCase) ||
+      lowerCase(column.description).includes(searchLowerCase) ||
+      lowerCase(getDataTypeString(column.dataType)).includes(searchLowerCase);
+
+    if (isContainData) {
+      return [...searchedCols, column];
+    } else if (!isUndefined(column.children)) {
+      const searchedChildren = searchInColumns<T>(
+        column.children as T[],
+        searchText
+      );
+      if (searchedChildren.length > 0) {
+        return [
+          ...searchedCols,
+          {
+            ...column,
+            children: searchedChildren,
+          },
+        ];
+      }
+    }
+
+    return searchedCols;
+  }, [] as T[]);
+
+  return searchedValue;
+};
